factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	long long int a;
+	long int a;
 	long int large;
 	long int b;
